pyproject.toml
[tool.poetry]
name = "audioldm-train"
version = "0.1.0"
description = ""
authors = ["haoheliu <867390095@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = ">=1.26.0,<2.0.0" 
ruamel-yaml = "^0.18.5"
wget = "^3.2"
kornia = "^0.7.0"
braceexpand = "^0.1.7"
webdataset = "^0.2.75"
hear21passt = {git = "https://github.com/haoheliu/passt_hear21.git"}
audioldm-eval = {git = "https://github.com/haoheliu/audioldm_eval.git"}
wandb = "^0.16.0"
pytorch-lightning = "^2.1.1"
einops = "^0.7.0"
h5py = "^3.10.0"
transformers = "4.30.2"
ftfy = "^6.1.1"
pandas = "^2.1.3"
matplotlib = "^3.8.1"
taming-transformers-rom1504 = "^0.0.6"
ipdb = "^0.13.13"
resampy = "0.4.1"
python-speech-features = "^0.6"
scipy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"